+++ crates/nu-protocol/src/engine/call.rs
@@ -186,8 +186,9 @@ impl Call<'_> {
         }
     }
 
-    /// Get the original AST expression for a positional argument. Does not usually work for IR
-    /// unless the decl specified `requires_ast_for_arguments()`
+    /// Get the original AST expression for a positional argument.
+    ///
+    /// Under IR this is usually `None`: argument AST is not retained for builtins.
     pub fn positional_nth<'a>(&'a self, stack: &'a Stack, index: usize) -> Option<&'a Expression> {
         match &self.inner {
             CallImpl::AstRef(call) => call.positional_iter().nth(index),
